changed datatype decimal to int

+++ TryCatchConsoleApp/TryCatchConsoleApp/Program.cs
@@ -17,27 +17,27 @@ namespace TryCatchConsoleApp
             try
             {
                 // List used by the program
-                List<decimal> numList = new List<decimal>() { 250m, 4000m, 2395m, 1044m, 24093m, 100000m };
+                List<int> numList = new List<int>() { 250, 4000, 2395, 1044, 24093, 100000 };
                 // Displays the list for the user. 
                 Console.WriteLine("Here is a list of whole numbers: ");
-                foreach (decimal n in numList)
+                foreach (int n in numList)
                 {
                     Console.WriteLine(n);
                 }
 
                 // Asks the user for a number
                 Console.WriteLine("Please enter a number to divide each number in the list by.");
-                decimal userNum = Convert.ToDecimal(Console.ReadLine());
+                int userNum = Convert.ToInt32(Console.ReadLine());
 
-                // Loop that takes each decimal in list,
+                // Loop that takes each int in list,
                 // Divides it by the number the user entered,
                 // And displays the result to screen
 
-                foreach (decimal num in numList)
+                foreach (int num in numList)
                 {
                     
                     // number in list divided by user number saved in i
-                    decimal i = num / userNum;
+                    int i = num / userNum;
                     Console.WriteLine(num + " divided by " + userNum + " equals " + i);
                 }
             }
